fixed progress bar
